Refactor directory structure
