+++ components/extensions/vscode/.vscode/extensions.json
