+++ components/extensions/vscode/.vscodeignore
