+++ components/extensions/vscode/resources/.gitignore
