+++ components/extensions/vscode/scripts/uninstall.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,18 +16,6 @@
  * under the License.
  */
 
-import * as vscode from "vscode";
-import ExtensionUtils from "./ExtensionUtils";
+import BalLangServerUtils from "../src/utils/BalLangServerUtils";
 
-export const activate = (context: vscode.ExtensionContext) => {
-    try {
-        ExtensionUtils.setupBalLangServerPlugins(context.extensionPath);
-    } catch (error) {
-        vscode.window.showErrorMessage("Failed to install Cellery code completion plugins");
-        throw error;
-    }
-};
-
-export const deactivate = () => {
-    // Do nothing
-};
+BalLangServerUtils.uninstallLangPlugins();
