+++ components/extensions/vscode/src/constants.ts
