+++ components/extensions/vscode/src/extension.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,6 +16,18 @@
  * under the License.
  */
 
-import BalLangServerUtils from "../src/BalLangServerUtils";
+import * as vscode from "vscode";
+import ExtensionUtils from "./utils/ExtensionUtils";
 
-BalLangServerUtils.uninstallLangPlugins();
+export const activate = (context: vscode.ExtensionContext) => {
+    try {
+        ExtensionUtils.setupBalLangServerPlugins(context.extensionPath);
+    } catch (error) {
+        vscode.window.showErrorMessage("Failed to install Cellery code completion plugins");
+        throw error;
+    }
+};
+
+export const deactivate = () => {
+    // Do nothing
+};
