+++ components/extensions/vscode/src/utils/CommonUtils.ts
