+++ components/extensions/vscode/test/suite/extension.test.ts
