+++ components/extensions/vscode/test/suite/index.ts
