+++ components/extensions/vscode/tslint.json
