+++ components/extensions/vscode/webpack.config.js
